Serial device implementation + USB device based on serial device
