+++ piksi_multi_cpp/CMakeLists.txt
@@ -11,21 +11,22 @@ catkin_simple(ALL_DEPS_REQUIRED)
 # LIBRARIES #
 #############
 cs_add_library(${PROJECT_NAME}
-  src/device/device.cc
-  src/device/device_factory.cc
-  src/device/device_usb.cc
-  src/device/device_tcp.cc
-  src/receiver/receiver_attitude.cc
-  src/receiver/receiver_base_station.cc
-  src/receiver/receiver_factory.cc
-  src/receiver/receiver_position.cc
-  src/receiver/receiver.cc
-  src/sbp_callback_handler/sbp_callback_handler.cc
-  src/sbp_callback_handler/sbp_callback_handler_factory.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_heartbeat.cc
-  src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_imu_raw.cc
-)
+        src/device/device.cc
+        src/device/device_factory.cc
+        src/device/device_usb.cc
+        src/device/device_serial.cc
+        src/device/device_tcp.cc
+        src/receiver/receiver_attitude.cc
+        src/receiver/receiver_base_station.cc
+        src/receiver/receiver_factory.cc
+        src/receiver/receiver_position.cc
+        src/receiver/receiver.cc
+        src/sbp_callback_handler/sbp_callback_handler.cc
+        src/sbp_callback_handler/sbp_callback_handler_factory.cc
+        src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay.cc
+        src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_heartbeat.cc
+        src/sbp_callback_handler/sbp_callback_handler_relay/sbp_callback_handler_relay_imu_raw.cc
+        )
 target_link_libraries(${PROJECT_NAME} serialport)
 
 ############
