+++ piksi_multi_cpp/include/piksi_multi_cpp/device/device_usb.h
@@ -2,27 +2,24 @@
 #define PIKSI_MULTI_CPP_DEVICE_DEVICE_USB_H_
 
 #include <libserialport.h>
+#include "device_serial.h"
 #include "piksi_multi_cpp/device/device.h"
 
 namespace piksi_multi_cpp {
 
-class DeviceUSB : public Device {
+class DeviceUSB : public DeviceSerial {
  public:
   DeviceUSB(const Identifier& id);
 
   // List all unique Piksi multi devices using the usb serial number.
   static Identifiers discoverAllSerialNumbers();
-  bool open() override;
-  int32_t read(uint8_t* buff, uint32_t n) const override;
-  void close() override;
 
  private:
   static Identifier identifyPiksiAndGetSerialNumber(struct sp_port* port);
   static void printDeviceInfo(struct sp_port* port);
   static void printPorts();
-  bool allocatePort();
+  bool allocatePort() override ;
 
-  struct sp_port* port_;
 };
 }  // namespace piksi_multi_cpp
