+++ piksi_multi_cpp/src/device/device_serial.cc
@@ -1,0 +1,116 @@
+#include "piksi_multi_cpp/device/device_serial.h"
+
+#include <libserialport.h>
+#include <ros/console.h>
+#include <regex>
+
+namespace piksi_multi_cpp {
+
+DeviceSerial::DeviceSerial(const Identifier& id) : Device(id), port_(nullptr) {}
+
+bool DeviceSerial::parseId() {
+  std::regex rgx("^(.+)@(\\d{4,7})$");
+
+  std::smatch match;
+  if (std::regex_search(id_, match, rgx) && match.size() > 1) {
+    portname_ = match.str(1);
+    std::stringstream(match.str(2)) >> baudrate_;
+    return true;
+  }
+  return false;
+}
+
+bool DeviceSerial::allocatePort() {
+  sp_return result = sp_get_port_by_name(portname_.c_str(), &port_);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("No serial port named " << portname_);
+    return false;
+  }
+
+  result = sp_set_baudrate(port_, baudrate_);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("Cannot set baud rate: " << result);
+    return false;
+  }
+  ROS_DEBUG("Configured the baud rate.");
+  return true;
+}
+
+bool DeviceSerial::open() {
+  if (!parseId()) {
+    return false;
+  }
+
+  // Allocate port.
+  if (!allocatePort()) {
+    close();
+    return false;
+  }
+
+  // Open port.
+  sp_return result = sp_open(port_, SP_MODE_READ);
+  if (result != SP_OK) {
+    ROS_ERROR("Cannot open port %s: %d", sp_get_port_name(port_), result);
+    close();
+    return false;
+  }
+  ROS_INFO_STREAM("Opened port: " << sp_get_port_name(port_));
+
+  // Configuration.
+  // https://github.com/swift-nav/libsbp/blob/master/c/example/example.c
+  result = sp_set_flowcontrol(port_, SP_FLOWCONTROL_NONE);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("Cannot set flow control: " << result);
+    close();
+    return false;
+  }
+  ROS_DEBUG("Configured the flow control.");
+
+  result = sp_set_bits(port_, 8);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("Cannot set data bits: " << result);
+    close();
+    return false;
+  }
+  ROS_DEBUG("Configured the number of data bits.");
+
+  result = sp_set_parity(port_, SP_PARITY_NONE);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("Cannot set parity: " << result);
+    close();
+    return false;
+  }
+  ROS_DEBUG("Configured the parity.");
+
+  result = sp_set_stopbits(port_, 1);
+  if (result != SP_OK) {
+    ROS_ERROR_STREAM("Cannot set stop bits: " << result);
+    close();
+    return false;
+  }
+  ROS_DEBUG("Configured the number of stop bits.");
+
+  return true;
+}
+
+int32_t DeviceSerial::read(uint8_t* buff, uint32_t n) const {
+  if (!port_) {
+    ROS_ERROR_STREAM("Port not opened.");
+    return 0;
+  }
+
+  return sp_blocking_read(port_, buff, n, 0);
+}
+
+void DeviceSerial::close() {
+  if (port_) {
+    sp_return result = sp_close(port_);
+    if (result != SP_OK) {
+      ROS_ERROR("Cannot close %s properly.", sp_get_port_name(port_));
+    }
+    sp_free_port(port_);
+    port_ = nullptr;
+  }
+}
+
+}  // namespace piksi_multi_cpp
