+++ piksi_multi_cpp/src/device/device_usb.cc
@@ -7,7 +7,7 @@ const int kInterfaceNumber = 1;
 
 namespace piksi_multi_cpp {
 
-DeviceUSB::DeviceUSB(const Identifier& id) : Device(id), port_(nullptr) {}
+DeviceUSB::DeviceUSB(const Identifier& id) : DeviceSerial(id) {}
 
 Identifiers DeviceUSB::discoverAllSerialNumbers() {
   Identifiers identifiers;
@@ -80,79 +80,6 @@ bool DeviceUSB::allocatePort() {
     ROS_INFO_STREAM(
         "Allocated USB port for Piksi Multi with serial number: " << id_);
     return true;
-  }
-}
-
-bool DeviceUSB::open() {
-  // Allocate port.
-  if (!allocatePort()) {
-    close();
-    return false;
-  }
-
-  // Open port.
-  sp_return result = sp_open(port_, SP_MODE_READ);
-  if (result != SP_OK) {
-    ROS_ERROR("Cannot open port %s: %d", sp_get_port_name(port_), result);
-    close();
-    return false;
-  }
-  ROS_INFO_STREAM("Opened port: " << sp_get_port_name(port_));
-
-  // Configuration.
-  // https://github.com/swift-nav/libsbp/blob/master/c/example/example.c
-  result = sp_set_flowcontrol(port_, SP_FLOWCONTROL_NONE);
-  if (result != SP_OK) {
-    ROS_ERROR_STREAM("Cannot set flow control: " << result);
-    close();
-    return false;
-  }
-  ROS_DEBUG("Configured the flow control.");
-
-  result = sp_set_bits(port_, 8);
-  if (result != SP_OK) {
-    ROS_ERROR_STREAM("Cannot set data bits: " << result);
-    close();
-    return false;
-  }
-  ROS_DEBUG("Configured the number of data bits.");
-
-  result = sp_set_parity(port_, SP_PARITY_NONE);
-  if (result != SP_OK) {
-    ROS_ERROR_STREAM("Cannot set parity: " << result);
-    close();
-    return false;
-  }
-  ROS_DEBUG("Configured the parity.");
-
-  result = sp_set_stopbits(port_, 1);
-  if (result != SP_OK) {
-    ROS_ERROR_STREAM("Cannot set stop bits: " << result);
-    close();
-    return false;
-  }
-  ROS_DEBUG("Configured the number of stop bits.");
-
-  return true;
-}
-
-int32_t DeviceUSB::read(uint8_t* buff, uint32_t n) const {
-  if (!port_) {
-    ROS_ERROR_STREAM("Port not opened.");
-    return 0;
-  }
-
-  return sp_blocking_read(port_, buff, n, 0);
-}
-
-void DeviceUSB::close() {
-  if (port_) {
-    sp_return result = sp_close(port_);
-    if (result != SP_OK) {
-      ROS_ERROR("Cannot close %s properly.", sp_get_port_name(port_));
-    }
-    sp_free_port(port_);
-    port_ = nullptr;
   }
 }
